Split up transformerlab-api-sdk into components
This change shouldn't change any functionality,
all it does is split the transformerlab-api-sdk into:
  * endpoints
  * functions
  * hooks

By doing so, later on we can clean things up by:
- reducing the number of functions
- automating the generartion of endpoints (which is
the most popular use for this sdk)
- creating a generic SWR hook for all endpoints
